statcol: Refactor for libti_rpmsg_char library
The ti_rpmsg_char library officially supports the rpmsg char
and replaces the older rpmsg helper libary.

It simplifies the process of finding the right remoteproc,
virtio, and directly allows to open the named endpoint
for the rpmsg_char device.

Update the code to use this new library
Few cosmetic updates to the rest of the code

Signed-off-by: Nikhil Devshatwar <nikhil.nd@ti.com>